requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
celery
argon2-cffi
psycopg2-binary
asyncpg
supabase
werkzeug
email_validator
orjson
redis
